requirements.txt
# Core dependencies for YOLO training
ultralytics>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
PyYAML>=5.4.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.3.0
tqdm>=4.62.0
tensorboard>=2.7.0

# Optional but recommended
psutil>=5.8.0
thop>=0.1.0

# For GPU support (if available)
# torch-audio>=0.9.0  # Uncomment if needed
